+++ src/main.js
@@ -1,5 +1,2 @@
 export default class ESToolbox {
-  constructor() {
-    this.something = "nothing";
-  }
 }